Pass testActivateAll test
