+++ sim-inventory/src/main/java/org/ostelco/SimAdministrationApplication.kt
@@ -10,23 +10,9 @@ import io.swagger.v3.oas.integration.SwaggerConfiguration
 import io.swagger.v3.oas.models.OpenAPI
 import io.swagger.v3.oas.models.info.Contact
 import io.swagger.v3.oas.models.info.Info
-import org.apache.commons.csv.CSVFormat
-import org.apache.commons.csv.CSVParser
 import org.hibernate.validator.constraints.NotEmpty
-import org.skife.jdbi.v2.StatementContext
-import org.skife.jdbi.v2.sqlobject.*
-import org.skife.jdbi.v2.sqlobject.customizers.BatchChunkSize
-import org.skife.jdbi.v2.sqlobject.customizers.RegisterMapper
-import org.skife.jdbi.v2.tweak.ResultSetMapper
-import java.io.BufferedReader
 import java.io.IOException
 import java.io.InputStream
-import java.io.InputStreamReader
-import java.nio.charset.Charset
-import java.sql.ResultSet
-import java.sql.SQLException
-import java.util.concurrent.ConcurrentLinkedDeque
-import java.util.concurrent.atomic.AtomicLong
 import java.util.stream.Collectors
 import java.util.stream.Stream
 import javax.ws.rs.*
@@ -118,84 +104,6 @@ class AppExceptionMapper : ExceptionMapper<Es2Exception> {
 }
 */
 
-/**
- *  Representing a single SIM card.
- */
-// @JsonSchema("SimEntry")
-data class SimEntry(
-        @JsonProperty("id") val id: Long? = null,
-        @JsonProperty("batch") val batch: Long,
-        @JsonProperty("hlrId") val hlrId: String,
-        @JsonProperty("smdpplus") val smdpplus: String? = null,
-        @JsonProperty("msisdn") val msisdn: String? = null,
-        @JsonProperty("iccid") val iccid: String,
-        @JsonProperty("imsi") val imsi: String,
-        @JsonProperty("eid") val eid: String? = null,
-        @JsonProperty("hlrActivation") val hlrActivation: Boolean = false,
-        @JsonProperty("smdpPlusActivation") val smdpPlusActivation: Boolean = false,
-        @JsonProperty("pin1") val pin1: String,
-        @JsonProperty("pin2") val pin2: String,
-        @JsonProperty("puk1") val puk1: String,
-        @JsonProperty("puk2") val puk2: String
-)
-
-
-/**
- * Describe a batch of SIM cards that was imported at some time
- */
-data class SimImportBatch(
-        @JsonProperty("id") val id: Long,
-        @JsonProperty("endedAt") val endedAt: Long,
-        @JsonProperty("message") val status: String?,
-        @JsonProperty("importer") val importer: String,
-        @JsonProperty("size") val size: Long,
-        @JsonProperty("hlr") val hlr: String,
-        @JsonProperty("profileVendor") val profileVendor: String
-)
-
-
-
-
-/**
- * An adapter that can connect to HLR entries and activate/deactivate
- * individual SIM profiles.
- */
-data class HlrAdapter(
-        @JsonProperty("id") val id: Long,
-        @JsonProperty("name") val name:String) {
-
-    /**
-     * Will connect to the HLR and then activate the profile, so that when
-     * a VLR asks the HLR for the an authentication triplet, then the HLR
-     * will know that it should give an answer.
-     */
-    fun activateEntry(simEntry: SimEntry) {
-        // XXX TBD
-    }
-}
-
-
-
-
-/**
- * An adapter that can connect to HLR entries and activate/deactivate
- * individual SIM profiles.
- */
-data class SmdpPlusAdapter(
-        @JsonProperty("id") val id: Long,
-        @JsonProperty("name") val name:String) {
-
-    /**
-     * Will connect to the SM-DP+  and then activate the profile, so that when
-     * user equpiment tries to download a profile, it will get a profile to
-     * download.
-     */
-    fun activateEntry(simEntry: SimEntry) {
-        // XXX TBD
-    }
-}
-
-
 ///
 ///  The web resource using the protocol domain model.
 ///
@@ -203,7 +111,7 @@ data class SmdpPlusAdapter(
 @Path("/ostelco/sim-inventory/{hlr}/")
 class EsimInventoryResource(val dao: SimInventoryDAO) {
 
-    private fun <T> assertNonNull(v: T?) : T {
+    private fun <T> assertNonNull(v: T?): T {
         if (v == null) {
             throw WebApplicationException(Response.Status.NOT_FOUND)
         } else {
@@ -219,7 +127,9 @@ class EsimInventoryResource(val dao: SimInventoryDAO) {
             @NotEmpty @PathParam("hlr") hlr: String,
             @NotEmpty @PathParam("iccid") iccid: String): SimEntry {
 
-        return assertNonNull(dao.getSimProfileByIccid(iccid))
+        val sim = assertNonNull(dao.getSimProfileByIccid(iccid))
+        assertHlrsEqual(hlr, sim.hlrId)
+        return sim
     }
 
     @Produces(MediaType.APPLICATION_JSON)
@@ -229,7 +139,9 @@ class EsimInventoryResource(val dao: SimInventoryDAO) {
     fun findByImsi(
             @NotEmpty @PathParam("hlr") hlr: String,
             @NotEmpty @PathParam("imsi") imsi: String): SimEntry {
-        return assertNonNull(dao.getSimProfileByImsi(imsi))
+        val sim = assertNonNull(dao.getSimProfileByImsi(imsi))
+        assertHlrsEqual(hlr, sim.hlrId)
+        return sim
     }
 
 
@@ -260,21 +172,20 @@ class EsimInventoryResource(val dao: SimInventoryDAO) {
     fun activateByIccid(
             @NotEmpty @PathParam("hlr") hlr: String,
             @NotEmpty @PathParam("iccid") iccid: String): SimEntry {
-        return SimEntry(
-                id = 1L,
-                msisdn = "82828282828282828",
-                hlrId = "foo",
-                batch = 99L,
-                iccid = iccid,
-                imsi = "9u8y32879329783247",
-                eid = "bb",
-                hlrActivation = false,
-                smdpPlusActivation = false,
-                pin1 = "ss",
-                pin2 = "ss",
-                puk1 = "ss",
-                puk2 = "ss"
-        )
+        val sim = activateHlrProfileByIccid(hlr, iccid)
+        assertHlrsEqual(hlr, sim.hlrId)
+        if (sim.smdpplus != null) {
+            return activateEsimProfileByIccid(hlr, iccid)
+        } else {
+            return sim
+        }
+    }
+
+    fun assertHlrsEqual(hlr1: String, hlr2: String) {
+        if (!hlr1.equals(hlr2)) {
+            // XXX log.error("Attempt to impersonate HLR.  '$hlr1', '$hlr2'")
+            throw WebApplicationException(Response.Status.BAD_REQUEST)
+        }
     }
 
     @Produces(MediaType.APPLICATION_JSON)
@@ -286,16 +197,16 @@ class EsimInventoryResource(val dao: SimInventoryDAO) {
             @NotEmpty @PathParam("iccid") iccid: String): SimEntry {
 
         val simEntry = assertNonNull(dao.getSimProfileByIccid(iccid))
+        assertHlrsEqual(hlr, simEntry.hlrId)
         val hlrAdapter = assertNonNull(dao.getHlrAdapterByName(hlr))
 
         try {
-            hlrAdapter.activateEntry(simEntry)
-        } catch (e:Exception) {
+            hlrAdapter.activate(simEntry)
+        } catch (e: Exception) {
             throw WebApplicationException(Response.Status.BAD_REQUEST)
         }
         return assertNonNull(dao.setActivatedInHlr(simEntry.id!!))
     }
-
 
     @Produces(MediaType.APPLICATION_JSON)
     @Consumes(MediaType.APPLICATION_JSON)
@@ -305,6 +216,7 @@ class EsimInventoryResource(val dao: SimInventoryDAO) {
             @NotEmpty @PathParam("hlr") hlr: String,
             @NotEmpty @PathParam("iccid") iccid: String): SimEntry {
         val simEntry = assertNonNull(dao.getSimProfileByIccid(iccid))
+        assertHlrsEqual(hlr, simEntry.hlrId)
 
         if (simEntry.smdpplus == null) {
             throw WebApplicationException(Response.Status.BAD_REQUEST)
@@ -326,21 +238,14 @@ class EsimInventoryResource(val dao: SimInventoryDAO) {
     fun deactrivateByIccid(
             @NotEmpty @PathParam("hlr") hlr: String,
             @NotEmpty @PathParam("iccid") iccid: String): SimEntry {
-        return SimEntry(
-                id = 1L,
-                msisdn = "2134234",
-                batch = 99L,
-                hlrId = "foo",
-                iccid = iccid,
-                imsi = "9u8y32879329783247",
-                eid = "bb",
-                hlrActivation = false,
-                smdpPlusActivation = false,
-                pin1 = "ss",
-                pin2 = "ss",
-                puk1 = "ss",
-                puk2 = "ss"
-        )
+        val simEntry = assertNonNull(dao.getSimProfileByIccid(iccid))
+
+        assertHlrsEqual(hlr, simEntry.hlrId)
+        val hlrAdapter = assertNonNull(dao.getHlrAdapterByName(hlr))
+
+        hlrAdapter.deactivate(simEntry)
+        dao.setActivatedInHlr(simEntry.id!!)
+        return dao.getSimProfileById(simEntry.id!!)
     }
 
     @Produces(MediaType.APPLICATION_JSON)
@@ -353,365 +258,56 @@ class EsimInventoryResource(val dao: SimInventoryDAO) {
             @NotEmpty @PathParam("profilevendor") profilevendor: String,
             csvInputStream: InputStream): SimImportBatch {
 
+        // XXX Check referential integrity of hlr and profilevendor
+        //      Also add a parameter for importer, and c heck if that importer
+        //      has the necessary permissions to perform imports.
+
         return dao.importSims(
                 importer = "importer",
                 hlr = hlr,
-                profileVendor = "Idemia",
+                profileVendor = profilevendor,
                 csvInputStream = csvInputStream)
     }
 }
 
-class SimEntryIterator(hlrId: String, batchId: Long, csvInputStream: InputStream) : Iterator<SimEntry> {
-
-    var count = AtomicLong(0)
-    // XXX The current implementation puts everything in a deque at startup.
-    //     This is correct, but inefficient, in partricular for large
-    //     batches.   Once proven to work, this thing should be rewritten
-    //     to use coroutines, to let the "next" get the next available
-    //     sim entry.  It may make sense to have a reader and writer thread
-    //     coordinating via the deque.
-    private val values = ConcurrentLinkedDeque<SimEntry>()
-
-    init {
-        // XXX Adjust to fit whatever format we should cater to, there may
-        //     be some variation between  sim vendors, and that should be
-        //     something we can adjust to given the parameters sent to the
-        //     reader class on creation.   Should anyway be configurable in
-        //     a config file or perhaps some config database.
-
-        val csvFileFormat = CSVFormat.DEFAULT
-                .withQuote(null)
-                .withFirstRecordAsHeader()
-                .withIgnoreEmptyLines(true)
-                .withTrim()
-                .withDelimiter(',')
-
-        BufferedReader(InputStreamReader(csvInputStream, Charset.forName(
-                "ISO-8859-1"))).use { reader ->
-            CSVParser(reader, csvFileFormat).use { csvParser ->
-                for (record in csvParser) {
-
-                    val iccid = record.get("ICCID")
-                    val imsi = record.get("IMSI")
-                    val pin1 = record.get("PIN1")
-                    val pin2 = record.get("PIN2")
-                    val puk1 = record.get("PUK1")
-                    val puk2 = record.get("PUK2")
-
-                    val value = SimEntry(
-                            batch = batchId,
-                            hlrId = hlrId,
-                            iccid = iccid,
-                            imsi = imsi,
-                            pin1 = pin1,
-                            puk1 = puk1,
-                            puk2 = puk2,
-                            pin2 = pin2
-                    )
-
-                    values.add(value)
-                    count.incrementAndGet()
-                }
-            }
-        }
-    }
-
-    /**
-     * Returns the next element in the iteration.
-     */
-    override operator fun next(): SimEntry {
-        return values.removeLast()
-    }
-
-    /**
-     * Returns `true` if the iteration has more elements.
-     */
-    override operator fun hasNext(): Boolean {
-        return !values.isEmpty()
-    }
-}
 
 /**
- * The DAO we're using to access the SIM inventory, and also the
- * pieces of SM-DP+/HLR infrastucture the SIM management needs to
- * be aware of.
+ * An adapter that can connect to HLR entries and activate/deactivate
+ * individual SIM profiles.
  */
-abstract class SimInventoryDAO {
-
-
-    @SqlQuery("select * from sim_entries where id = :id")
-    @RegisterMapper(SimEntryMapper::class)
-    abstract fun getSimProfileById(@Bind("id")id: Long): SimEntry
-
-    @SqlQuery("select * from sim_entries where iccid = :iccid")
-    @RegisterMapper(SimEntryMapper::class)
-    abstract fun getSimProfileByIccid(@Bind("iccid")iccid: String): SimEntry
-
-    @SqlQuery("select * from sim_entries where imsi = :imsi")
-    @RegisterMapper(SimEntryMapper::class)
-    abstract fun getSimProfileByImsi(@Bind("imsi")imsi: String): SimEntry
-
-    @SqlQuery("select * from sim_entries where msisdn = :msisdn")
-    @RegisterMapper(SimEntryMapper::class)
-    abstract fun getSimProfileByMsisdn(@Bind("msisdn")msisdn: String): SimEntry
-
-
-    class SimEntryMapper : ResultSetMapper<SimEntry> {
-
-        @Throws(SQLException::class)
-        override fun map(index: Int, r: ResultSet, ctx: StatementContext): SimEntry? {
-            if (r.isAfterLast) {
-                return null
-            }
-
-            val id = r.getLong("id")
-            val batch = r.getLong("batch")
-            val hlrId = r.getString("hlrId")
-            val smdpplus = r.getString("smdpplus")
-            val msisdn = r.getString("msisdn")
-            val iccid = r.getString("iccid")
-            val imsi = r.getString("imsi")
-            val eid = r.getString("eid")
-            val smdpPlusActivation = r.getBoolean("smdpPlusActivation")
-            val hlrActivation = r.getBoolean("hlrActivation")
-            val pin1 = r.getString("pin1")
-            val pin2 = r.getString("pin2")
-            val puk1 = r.getString("puk1")
-            val puk2 = r.getString("puk2")
-
-            return SimEntry(
-                    id = id,
-                    batch = batch,
-                    hlrId = hlrId,
-                    smdpplus = smdpplus,
-                    msisdn = msisdn,
-                    iccid = iccid,
-                    imsi = imsi,
-                    eid = eid,
-                    smdpPlusActivation = smdpPlusActivation,
-                    hlrActivation = hlrActivation,
-                    pin1 = pin1,
-                    pin2 = pin2,
-                    puk1 = puk1,
-                    puk2 = puk2
-            )
-        }
-    }
-
-    @SqlQuery("select * from hlr_adapters where name = :name")
-    @RegisterMapper(HlrAdapterMapper::class)
-    abstract fun getHlrAdapterByName(@Bind("name")name: String): HlrAdapter
-
-
-    @SqlQuery("select * from hlr_adapters where id = :id")
-    @RegisterMapper(HlrAdapterMapper::class)
-    abstract fun getHlrAdapterByName(@Bind("id")id: Long): HlrAdapter
-
-
-    class HlrAdapterMapper : ResultSetMapper<HlrAdapter> {
-
-        @Throws(SQLException::class)
-        override fun map(index: Int, r: ResultSet, ctx: StatementContext): HlrAdapter? {
-            if (r.isAfterLast) {
-                return null
-            }
-
-            val id = r.getLong("id")
-            val name = r.getString("name")
-
-            return HlrAdapter(
-                    id = id,
-                    name = name)
-        }
-    }
-
-    abstract fun getSmdpPlusAdapterByName(name: String): SmdpPlusAdapter?
-
-    //
-    // Getting the ID of the last insert, regardless of table
-    //
-    @SqlQuery("select last_insert_rowid()")
-    abstract fun lastInsertRowid(): Long
-
-    //
-    // Importing
-    //
-    @Transaction
-    @SqlBatch("INSERT INTO sim_entries (iccid, imsi, pin1, pin2, puk1, puk2) VALUES (:iccid, :imsi, :pin1, :pin2, :puk1, :puk2)")
-    @BatchChunkSize(1000)
-    abstract fun insertAll(@BindBean entries: Iterator<SimEntry>);
-
-
-    @SqlUpdate("INSERT INTO sim_import_batches (status,  importer, hlr, profileVendor) VALUES ('STARTED', :importer, :hlr, :profileVendor)")
-    abstract fun createNewSimImportBatch(
-            @Bind("importer") importer: String,
-            @Bind("hlr") hlr: String,
-            @Bind("profileVendor") profileVendor: String)
-
-    abstract fun getIdOfBatchCreatedLast(): Long
-
-    @SqlUpdate("UPDATE sim_import_batches SET size = :size, status = :status, endedAt = :endedAt  WHERE id = :id")
-    abstract fun updateBatchState(
-            @Bind("id") id: Long,
-            @Bind("size") size: Long,
-            @Bind("status") status: String,
-            @Bind("endedAt") endedAt: Long)
-
-    @Transaction
-    fun importSims(
-            importer: String,
-            hlr: String,
-            profileVendor: String,
-            csvInputStream: InputStream): SimImportBatch {
-
-        createNewSimImportBatch(
-                importer = importer,
-                hlr = hlr,
-                profileVendor = profileVendor)
-        val batchId = lastInsertRowid()
-        val values = SimEntryIterator(
-                hlrId = hlr,
-                batchId = batchId,
-                csvInputStream = csvInputStream)
-        insertAll(values)
-        updateBatchState(
-                id = batchId,
-                size = values.count.get(),
-                status = "SUCCESS",
-                endedAt = System.currentTimeMillis())
-        return getBatchInfo(batchId)
-    }
-
-    @SqlQuery("select * from sim_import_batches where id = :id")
-    @RegisterMapper(SimImportBatchMapper::class)
-    abstract fun getBatchInfo(@Bind("id") id: Long): SimImportBatch
-
-    class SimImportBatchMapper : ResultSetMapper<SimImportBatch> {
-
-        @Throws(SQLException::class)
-        override fun map(index: Int, r: ResultSet, ctx: StatementContext): SimImportBatch? {
-            if (r.isAfterLast) {
-                return null
-            }
-
-            val id = r.getLong("id")
-            val endedAt = r.getLong("endedAt")
-            val status = r.getString("status")
-            val hlr = r.getString("hlr")
-            val profileVendor = r.getString("profileVendor")
-            val size = r.getLong("size")
-
-            return SimImportBatch(
-                    id = id,
-                    endedAt = endedAt,
-                    status = status,
-                    hlr = hlr,
-                    profileVendor = profileVendor,
-                    size = size,
-                    importer = "XXX Replace with name of agent that facilitated the import")
-        }
-    }
-
-    //
-    // Setting activation statuses
-    //
-
-    @SqlUpdate("UPDATE sim_entries SET hlrActivation = :hlrActivation  WHERE id = :id")
-    abstract fun setHlrActivation(
-            @Bind("id") id:Long,
-            @Bind("hlrActivation") hlrActivation:Boolean)
-
-
-    @SqlUpdate("UPDATE sim_entries SET smdpPlusActivation = :smdpPlusActivation  WHERE id = :id")
-    abstract fun setSmdpPlusActivation(
-            @Bind("id") id:Long,
-            @Bind("smdpPlusActivation") smdpPlusActivation:Boolean)
-
+data class HlrAdapter(
+        @JsonProperty("id") val id: Long,
+        @JsonProperty("name") val name: String) {
 
     /**
-     * Set the entity to be marked as "active" in the HLR, then return the
-     * SIM entry.
+     * Will connect to the HLR and then activate the profile, so that when
+     * a VLR asks the HLR for the an authentication triplet, then the HLR
+     * will know that it should give an answer.
      */
-    fun setActivatedInHlr(id: Long): SimEntry? {
-        setHlrActivation(id, true)
-        return getSimProfileById(id)
+    fun activate(simEntry: SimEntry) {
+        // XXX TBD
     }
 
-
-    fun setActivatedInSmdpPlus(id: Long): SimEntry? {
-        setSmdpPlusActivation(id, true)
-        return getSimProfileById(id)
+    fun deactivate(simEntry: SimEntry) {
+        // XXX TBD
     }
+}
 
-    //
-    //  Binding a SIM card to an MSISDN
-    //
 
-    @SqlUpdate("UPDATE sim_entries SET msisdn = :msisdn  WHERE id = :id")
-    abstract fun setMsisdnOfSim(@Bind("id") id:Long, @Bind("msisdn") msisdn:String)
+/**
+ * An adapter that can connect to HLR entries and activate/deactivate
+ * individual SIM profiles.
+ */
+data class SmdpPlusAdapter(
+        @JsonProperty("id") val id: Long,
+        @JsonProperty("name") val name: String) {
 
-    //
-    // Finding next free SIM card for a particular HLR.
-    //
-    @SqlQuery("SELECT * FROM sim_entries WHERE hlr = :hlr AND msisdn = null limit 1")
-    @RegisterMapper(SimImportBatchMapper::class)
-    abstract fun findNextFreeSimForMsisdn(@Bind("hlr") hlr:String): SimEntry
-
-    //
-    // Allocating next free simcards in an HLR.
-    //
-    @Transaction
-    fun allocateNextFreeSimForMsisdn(hlr:String, msisdn: String): SimEntry? {
-        val sim : SimEntry= findNextFreeSimForMsisdn(hlr)
-        if (sim == null) {
-            return null
-        }
-
-        setMsisdnOfSim(sim.id!!, msisdn)
-
-        // This is an inefficient way of getting an updated profile,
-        // but it will work, and we can optimize it away if the need ever
-        // arises.
-        return getSimProfileByMsisdn(msisdn)
+    /**
+     * Will connect to the SM-DP+  and then activate the profile, so that when
+     * user equpiment tries to download a profile, it will get a profile to
+     * download.
+     */
+    fun activateEntry(simEntry: SimEntry) {
+        // XXX TBD
     }
-
-
-
-
-
-
-    //
-    // Creating and deleting tables (XXX only used for testing, and should be moved to
-    // a test only DAO eventually)
-    //
-    @SqlUpdate("create table sim_import_batches (id integer primary key autoincrement, status text, endedAt integer, importer text, size integer, hlr text, profileVendor text)")
-    abstract fun createImportBatchesTable();
-
-    @SqlUpdate("drop  table sim_import_batches")
-    abstract fun dropImportBatchesTable();
-
-
-    @SqlUpdate("create table sim_entries (id integer primary key autoincrement, hlrid text, smdpplus text, msisdn text, eid text, hlrActivation boolean, smdpPlusActivation boolean, batch integer, imsi varchar(15), iccid varchar(22), pin1 varchar(4), pin2 varchar(4), puk1 varchar(80), puk2 varchar(80), CONSTRAINT Unique_Imsi UNIQUE (imsi), CONSTRAINT Unique_Iccid UNIQUE (iccid))")
-    abstract fun createSimEntryTable();
-
-    @SqlUpdate("drop  table sim_entries")
-    abstract fun dropSimEntryTable();
-
-
-    @SqlUpdate("create table hlr_adapters (id integer primary key autoincrement, name text,  CONSTRAINT Unique_Name UNIQUE (name))")
-    abstract fun createHlrAdapterTable();
-
-    @SqlUpdate("drop  table hlr_adapters")
-    abstract fun dropHlrAdapterTable();
-
-
-    @SqlUpdate("create table smdp_plus_adapters (id integer primary key autoincrement, name text,  CONSTRAINT Unique_Name UNIQUE (name))")
-    abstract fun createSmdpPlusTable();
-
-    @SqlUpdate("drop  table smdp_plus_adapters")
-    abstract fun dropSmdpPlusTable();
-
-
-
-
 }
